Improved form generation/parsing
Added templates for form generation. Added the ability to read URL
parameters. Added a function to read form inputs. The reading
functions all return objects with properties and values that
correspond to the user's input.

Currently has support for select, numbers, strings, and checkboxes.
Conversion functions and default values can be specified with the
form.